Modify code of the displayed source code, so it is in a form of a standalone ScalaFX application.
This is turning a code of a _class_ (that derives from EnsembleExample) into an _object_ (JFXApp). In some situations it may lead to initialization issues. For instance, is a variable is defined at the end of the class it may not be properly initialized in the app, resulting in NPE. That variable may need to be copied before `getContent` method (in the original code) or turned into a function (replace `var` with `def`). All examples need to be tested/inspected to ensure that they work after turning into an app.